Crate gtypes

Fundamental types for GLib and GObject introspection.

This crate provides definitions for types that are intrinsic in GObject introspection, unlike all other types that have a definition in a GIR file. All crates generated from GObject introspection data should use these shared definitions.

Some types are omitted because a suitable equivalent is readily available in Rust:

Fixed-size integer types. These have straightforward built-in counterparts in Rust.

Strings can have GI types utf8 or filename. In the FFI all string values are represented as raw pointers to gchar. A safe representation of C strings in idiomatic Rust bindings needs some wrapper types which are not defined here.

Reexports

pub use primitive::{gboolean, gchar, guchar, gshort, gushort, gint, guint};

pub use primitive::{glong, gulong, gsize, gssize, gintptr, guintptr};

pub use primitive::{gfloat, gdouble, gunichar, gpointer, gconstpointer};

pub use primitive::{FALSE, TRUE};

Modules

primitive
Definitions for primitive fundamental types.

Type Definitions

GType
An integer type that designates GObject type identifiers.