Module gtypes::primitive

Definitions for primitive fundamental types.

Values of these types don't have associated semantics or state that would necessitate some kind of wrapper types in safe Rust bindings. Therefore, these types can be used in both FFI declarations and idiomatic Rust bindings, unlikely to require name disambiguation with any generated or binding types.

The untyped pointer types are also defined in this module, because safe Rust bindings are unlikely to need to reuse their names.

Constants

FALSE

TRUE

Type Definitions

gboolean

gchar

gconstpointer

gdouble

gfloat

gint

gintptr

glong

gpointer

gshort

gsize

gssize

guchar

guint

guintptr

gulong

gunichar

gushort